Move Ident code to own file
